Fix bug where user hitting cancel was returning 0
Fix bug where when user hits cancel or ok button without input,
the program would return that variable as a 0 instead of null, leading
to further issues. Now when user enters 0 in prompt, they will
be notified the number is out of range. And when the user hits cancel,
the variable will be saved as "null" instead of 0.